add history vc

+++ Heart Rate Monitor/Heart Rate Monitor/App/Extension/Date+.swift
@@ -16,8 +16,8 @@ extension Date {
         return formatter.string(from: self)
     }
     
-    func fromTimeMilli(timeMilli: String) -> Date{
-        return NSDate(timeIntervalSince1970: TimeInterval(Int(timeMilli)!)) as Date
+    func fromTimeMilli(timeMilli: String) -> Date {
+        return NSDate(timeIntervalSince1970: TimeInterval(Double(timeMilli)!)) as Date
     }
     
     var hour: Int {
